pyproject.toml
[tool.poetry]
name = "dtcli"
version = "0.1.0"
description = ""
authors = ["O-J1 <18110006+O-J1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pick = "^2.2.0"
tomli-w = "^1.0.0"
click = "^8.1.7"
pillow = "^10.0.0"

[tool.poetry.scripts]
dtcli = "dtcli.cli:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
